fix: CozoDB query injection, deduplicate aliases & debounce (#286, #284, #283)
- Escape all user-controlled strings in CozoDB queries via escapeForCozo()
- Remove duplicate escapeForCozo() from doc-indexer.ts, import from cozo-utils
- Hoist FEMALE_ALIASES, MALE_ALIASES, getRandomAlias to lib/agent-utils.ts
- Replace inline debounce in useTerminal.ts and immersive/page.tsx with import from lib/utils
- Remove unused throttle() export from lib/utils.ts
- Bump version to 0.25.2

Closes #286, closes #284, closes #283
